syntropism/infra/llm_proxy.py
"""
LLM Proxy Service Module

This module implements the LLM Proxy service that:
- Exposes a /llm endpoint for routing requests
- Enforces token quotas
- Logs all interactions
"""


from fastapi import APIRouter, HTTPException, Request
from loguru import logger
from pydantic import BaseModel

router = APIRouter()

# Token quota tracking (in-memory for stub implementation)
token_quotas = {}


class LLMRequest(BaseModel):
    prompt: str
    model: str
    max_tokens: int | None = 1000


class LLMResponse(BaseModel):
    response: str
    tokens_used: int
    model: str


@router.post("/llm")
async def handle_llm_request(request: LLMRequest, req: Request):
    """
    Handle LLM proxy requests with token quota enforcement and logging.
    """
    client_id = req.client.host if req.client else "unknown"

    logger.info(f"[component:llm_proxy] LLM request received from {client_id} for model {request.model}")

    # Check and enforce token quotas
    current_usage = token_quotas.get(client_id, 0)
    requested_tokens = request.max_tokens or 1000

    if current_usage + requested_tokens > 10000:  # Example quota limit
        logger.warning(f"[component:llm_proxy] Token quota exceeded for client {client_id}")
        raise HTTPException(status_code=429, detail="Token quota exceeded")

    # Update token usage
    token_quotas[client_id] = current_usage + requested_tokens

    # Log interaction
    logger.debug(f"[component:llm_proxy] LLM request - Client: {client_id}, Model: {request.model}, Tokens: {requested_tokens}")

    # Stub implementation - in production, this would route to actual LLM provider
    response_text = f"Stub response for prompt: {request.prompt}"

    logger.info(f"[component:llm_proxy] LLM response generated for {client_id}, tokens used: {requested_tokens}")

    return LLMResponse(
        response=response_text,
        tokens_used=requested_tokens,
        model=request.model
    )


@router.get("/llm/quota/{client_id}")
async def get_quota(client_id: str):
    """
    Get current token quota usage for a client.
    """
    usage = token_quotas.get(client_id, 0)
    return {"client_id": client_id, "tokens_used": usage, "quota_limit": 10000}


@router.post("/llm/quota/reset/{client_id}")
async def reset_quota(client_id: str):
    """
    Reset token quota for a client (admin operation).
    """
    token_quotas[client_id] = 0
    logger.info(f"[component:llm_proxy] Token quota reset for client {client_id}")
    return {"status": "quota_reset", "client_id": client_id}
