pyproject.toml
[project]
name = "syntropism"
version = "0.1.0"
description = "Economically motivated self-modifying agents"
authors = [
    {name = "ben-pollard",email = "benpollard0@googlemail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "langchain (>=1.2.9,<2.0.0)",
    "langgraph (>=1.0.8,<2.0.0)",
    "deepagents (>=0.3.12,<0.4.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "fastapi (>=0.115.0,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "docker (>=7.1.0,<8.0.0)",
    "loguru (>=0.7.2,<0.8.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "nats-py (>=2.13.1,<3.0.0)",
    "opentelemetry-sdk (>=1.39.1,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.39.1,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.poetry.scripts]
syntropism = "syntropism.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


# Ruff configuration - fast Python linter (replaces flake8, isort, etc.)
[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py


[tool.ruff.lint.pydocstyle]
# Enable docstring checking
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# MyPy configuration - static type checker
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter typing

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=syntropism --cov-report=term-missing"
markers = [
    "unit: Fast isolated tests",
    "integration: Tests requiring database or API",
    "e2e: Full system end-to-end tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::ResourceWarning"
]